[Chore #153129811] Add tests for App.js
