+++ package.json
@@ -22,6 +22,7 @@
     "redux-form": "^7.1.2",
     "redux-form-material-ui": "^4.3.0",
     "redux-logger": "^3.0.6",
+    "redux-mock-store": "^1.4.0",
     "redux-promise": "^0.5.3",
     "redux-thunk": "^2.2.0"
   },
@@ -40,6 +41,7 @@
     "eslint-plugin-react": "^7.4.0",
     "jest-localstorage-mock": "^2.1.0",
     "prettier-eslint": "^8.2.1",
-    "redux-devtools-extension": "^2.13.2"
+    "redux-devtools-extension": "^2.13.2",
+    "redux-test-utils": "^0.2.2"
   }
 }
